refactor(drizzle): tableFromType returns the slim table typed T from the dialect index
