+++ packages/drizzle-orm-pg-core/src/index.ts
@@ -23,7 +23,7 @@
 export * from './columns.ts';
 
 // The table factories and schema handles.
-export {pgTable, pgTableCreator, pgSchema} from './table.ts';
+export {pgTable, pgTableCreator, pgSchema, tableFromType} from './table.ts';
 export type {PgExtraConfigColumns, PgExtraConfigFn, PgSchema, PgSequence, PgSequenceOptions, PgTable} from './table.ts';
 
 // The pure-types road: the modifier markers applicable to pg columns. The
